+++ src/renderer/components/stages/Certificates.tsx
@@ -25,7 +25,7 @@ import ProgressCard from "../common/ProgressCard";
 import { createTheme } from '@mui/material/styles';
 import { stages } from "../configuration-wizard/Wizard";
 import { setActiveStep } from "./progress/activeStepSlice";
-import { getStageDetails, getSubStageDetails } from "./progress/progressStore";
+import { getStageDetails, getSubStageDetails } from "../../../utils/StageDetails";
 
 const Certificates = () => {
