+++ src/renderer/components/stages/Planning.tsx
@@ -29,7 +29,7 @@ import { alertEmitter } from "../Header";
 import { Checkbox, FormControlLabel } from "@mui/material";
 import { setActiveStep } from './progress/activeStepSlice';
 import EditorDialog from "../common/EditorDialog";
-import { getStageDetails } from "../stages/progress/progressStore";
+import { getStageDetails } from "../../../utils/StageDetails";
 
 // TODO: Our current theoretical cap is 72 (possibly minus a couple for "\n", 70?) But we force more chars in InstallationHandler.tsx
 // This is all I want to manually test for now. Future work can min/max this harder
