+++ src/renderer/components/stages/Security.tsx
@@ -26,7 +26,7 @@ import { createTheme } from '@mui/material/styles';
 import progressSlice from "./progress/progressSlice";
 import { stages } from "../configuration-wizard/Wizard";
 import { setActiveStep } from "./progress/activeStepSlice";
-import { getStageDetails, getSubStageDetails } from "./progress/progressStore";
+import { getStageDetails, getSubStageDetails } from "../../../utils/StageDetails";
 
 const Security = () => {
   const theme = createTheme();
