+++ src/renderer/components/stages/progress/StageProgressStatus.ts
