+++ src/renderer/components/stages/progress/activeStepSlice.ts
@@ -10,7 +10,7 @@
 
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { RootState } from '../../../store';
-import { setActiveStage } from './progressStore'
+import { setActiveStage } from './StageProgressStatus'
 
 export interface activeStep {
   activeStepIndex: number,
